WIP ...requested changes in code refactoring
